+++ src/TinyGsmClientSaraR4.h
@@ -240,7 +240,7 @@ class TinyGsmSaraR4
      * channel number after a successful connection.
      */
     explicit GsmClientSaraR4(TinyGsmSaraR4& modem, uint8_t /*mux*/ = 0) {
-      init(&modem, -1);
+      init(&modem, static_cast<uint8_t>(-1));
       is_secure = false;
     }
